+++ lib/plugins/index.js
@@ -8,7 +8,7 @@ module.exports = [
   require('./package/package.js'),
   require('./deploy.js'),
   require('./invoke/invoke.js'),
-  require('./info/info.js'),
+  require('./info.js'),
   require('./logs/logs.js'),
   require('./metrics/metrics.js'),
   require('./print/print.js'),
