+++ lib/plugins/info.js
